refactor(halley-wl): split compositor subsystems and tighten drag edge panning
- Move `activity` and `debug_scene` under `compositor` and update focus, fullscreen, clusters, workspace, and runtime to use the new module layout.
- Replace `compositor::surface_ops` with a dedicated `compositor::surface` module split into query and XDG helpers, and route backend, window, IPC, input, focus, and Wayland handlers through it.
- Break up large compositor files by extracting screenshot state/region helpers, overlap carry/resolve logic, and spawn reveal placement into smaller subsystem-owned modules.
- Tighten drag edge panning so it only activates for fully visible active surfaces and reduce the pan speed to make drag-driven viewport movement less aggressive.
- Add targeted tests for drag edge-pan eligibility, stacking surface helpers, overlap resolution behavior, and the relocated debug scene utilities.
