+++ crates/halley-wl/src/compositor/fullscreen/mod.rs
@@ -11,11 +11,11 @@ use smithay::reexports::wayland_server::{
     Resource, backend::ObjectId, protocol::wl_surface::WlSurface,
 };
 
-use crate::activity::{CommitActivity, VisualState};
 use crate::animation::{AnimSpec, AnimStyle};
+use crate::compositor::activity::{CommitActivity, VisualState};
 use crate::compositor::ctx::FullscreenCtx;
+use crate::compositor::debug_scene::{DebugScene, build_debug_scene};
 use crate::compositor::root::Halley;
-use crate::debug_scene::{DebugScene, build_debug_scene};
 
 pub mod state;
 pub mod system;
