+++ crates/halley-wl/src/compositor/surface/mod.rs
@@ -1,0 +1,15 @@
+mod query;
+mod xdg;
+
+#[cfg(test)]
+pub(crate) use query::stacking_render_order_map;
+pub(crate) use query::{
+    active_stacking_front_member_for_monitor, active_stacking_render_order_for_monitor,
+    active_stacking_visible_members_for_monitor, is_active_cluster_workspace_member,
+    is_active_stacking_workspace_member, node_allows_interactive_resize,
+    stack_focus_target_for_node,
+};
+pub(crate) use xdg::{
+    current_surface_size_for_node, request_close_focused_toplevel, request_close_node_toplevel,
+    request_toplevel_resize_mode, toplevel_min_size_for_node, window_geometry_for_node,
+};
