+++ crates/halley-wl/src/compositor/workspace/mod.rs
@@ -11,7 +11,7 @@ use smithay::reexports::wayland_server::{
 use smithay::wayland::compositor::with_states;
 use smithay::wayland::shell::xdg::XdgToplevelSurfaceData;
 
-use crate::activity::CommitActivity;
+use crate::compositor::activity::CommitActivity;
 use crate::compositor::ctx::{SurfaceLifecycleCtx, WorkspaceCtx};
 use crate::compositor::root::Halley;
